Move likes and stocks related code to a new file
